Revert "feat(suite-native): Display sell / exchange form even with view-only wallet"
This reverts commit c8e4aa82
