+++ suite-native/app/e2e/tests/tradingExchangeFlow.test.ts
@@ -4,8 +4,6 @@ import { MNEMONICS } from '@trezor/trezor-user-env-link';
 import { ethCoinEnabled } from '../fixtures/ethCoinEnabled';
 import { onboardingCompletedState } from '../fixtures/onboardingCompletedState';
 import { portfolioTrackerBtcAccountState } from '../fixtures/portfolioTrackerBtcAccountState';
-import { onDeviceConnecting } from '../pageObjects/deviceConnectingActions';
-import { onHome } from '../pageObjects/homeActions';
 import { onPassphrase } from '../pageObjects/passphraseModule';
 import { onTabBar } from '../pageObjects/tabBarActions';
 import { exchangePreviewActions } from '../pageObjects/trading/exchangePreviewActions';
@@ -38,52 +36,6 @@ conditionalDescribe(device.getPlatform() === 'android', 'Trade Exchange', () => 
 
         it('should display info card', async () => {
             await tradingExchangeActions.expectPortfolioTrackerInfoCard();
-        });
-    });
-
-    conditionalDescribe(isCIRun || passphrase, 'with device disconnected [@fixT3W1]', () => {
-        beforeAll(() => {
-            if (!passphrase) {
-                throw new Error(
-                    'TRADING_ACADEMIC_SEED_WALLET_PASSPHRASE environment variable is required',
-                );
-            }
-        });
-
-        beforeEach(async () => {
-            await openApp({ args: { preloadedState: preloadedStateWithTrezor } });
-            await prepareTrezorEmulator({
-                seed: MNEMONICS.mnemonic_academic,
-                passphrase_protection: true,
-            });
-            await waitForVisible(by.text('Connected'));
-            await onPassphrase.openPassphraseWallet(passphrase);
-            await onHome.waitForScreen();
-            await onDeviceConnecting.stopEmuAndConfirmViewOnlyWarning();
-            await tradingExchangeActions.openSwapForm();
-        });
-
-        it('should request trezor connect before review', async () => {
-            await tradingExchangeActions.selectSendAsset('USDC');
-            await tradingExchangeActions.selectReceiveAsset('USDT', 'Ethereum');
-            await tradingExchangeActions.selectReceiveAccount('Ethereum #1');
-            await tradingExchangeActions.setSendCryptoAmount('10');
-            await tradingExchangeActions.waitForQuotesToLoad();
-
-            await tradingExchangeActions.scrollScreenToBottom();
-            await tradingExchangeActions.expectValidExchangeForm();
-
-            await tradingExchangeActions.confirmTradingForm();
-
-            await exchangePreviewActions.expectExchangePreviewScreenToBeVisible();
-            await exchangePreviewActions.waitForFeesToLoad();
-            await exchangePreviewActions.scrollScreenToBottom();
-            await exchangePreviewActions.goToTransactionSigning();
-
-            await exchangeOutputsReviewActions.expectConnectTrezorInfo();
-            await exchangeOutputsReviewActions.cancelConnectTrezorInfo();
-
-            await tradingExchangeActions.waitForTradeDataToLoad();
         });
     });
