+++ suite-native/module-trading/package.json
@@ -41,7 +41,6 @@
         "@suite-native/atoms": "workspace:*",
         "@suite-native/clipboard": "workspace:*",
         "@suite-native/device": "workspace:*",
-        "@suite-native/device-authorization": "workspace:*",
         "@suite-native/device-manager": "workspace:*",
         "@suite-native/device-mutex": "workspace:*",
         "@suite-native/discovery": "workspace:*",
