+++ suite-native/module-trading/src/components/exchange/__tests__/ExchangeTab.comp.test.tsx
@@ -70,10 +70,10 @@ describe('ExchangeTab', () => {
         expect(queryByText('View-only wallet')).toBeNull();
     });
 
-    it('should display exchange form for view-only wallet', async () => {
+    it('should display View-only info with view-only wallet', async () => {
         mockIsDeviceInViewOnlyMode = true;
         const { getByText } = await renderExchangeTab();
 
-        expect(getByText('You pay')).toBeOnTheScreen();
+        expect(getByText('View-only wallet')).toBeOnTheScreen();
     });
 });
