+++ suite-native/module-trading/src/components/sell/__tests__/SellTab.comp.test.tsx
@@ -1,5 +1,5 @@
 import { FeatureFlag } from '@suite-native/feature-flags';
-import { PreloadedState, act, renderWithStoreProviderAsync } from '@suite-native/test-utils';
+import { PreloadedState, renderWithStoreProviderAsync } from '@suite-native/test-utils';
 
 import { SellTab } from '../SellTab';
 
@@ -21,14 +21,8 @@ jest.mock('../../../hooks/sell/useSellData', () => ({
 }));
 
 describe('SellTab', () => {
-    const renderSellTab = async (preloadedState: PreloadedState = {}) => {
-        const result = await renderWithStoreProviderAsync(<SellTab />, { preloadedState });
-
-        // wait for form reactions to run
-        await act(() => Promise.resolve());
-
-        return result;
-    };
+    const renderSellTab = (preloadedState: PreloadedState = {}) =>
+        renderWithStoreProviderAsync(<SellTab />, { preloadedState });
 
     beforeEach(() => {
         mockIsDeviceInViewOnlyMode = false;
@@ -84,7 +78,7 @@ describe('SellTab', () => {
         expect(queryByText('View-only wallet')).toBeNull();
     });
 
-    it('should display form even with view-only wallet', async () => {
+    it('should display View-only info with view-only wallet', async () => {
         mockIsDeviceInViewOnlyMode = true;
         const { getByText } = await renderSellTab({
             featureFlags: {
@@ -92,16 +86,6 @@ describe('SellTab', () => {
             },
         });
 
-        expect(getByText('Select asset')).toBeOnTheScreen();
-    });
-
-    it('should display form otherwise', async () => {
-        const { getByText } = await renderSellTab({
-            featureFlags: {
-                [FeatureFlag.IsTradingSellEnabled]: true,
-            },
-        });
-
-        expect(getByText('Select asset')).toBeOnTheScreen();
+        expect(getByText('View-only wallet')).toBeOnTheScreen();
     });
 });
