+++ suite-native/module-trading/tsconfig.json
@@ -49,7 +49,6 @@
         { "path": "../atoms" },
         { "path": "../clipboard" },
         { "path": "../device" },
-        { "path": "../device-authorization" },
         { "path": "../device-manager" },
         { "path": "../device-mutex" },
         { "path": "../discovery" },
